google-docs-test/requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
